[Breaking change] Remove obsolete TraceHelper
