--- Src/zipkin4net/Src/Utils/TraceHelper.cs
@@ -1,71 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace zipkin4net.Utils
-{
-    [System.Obsolete("TraceHelper is deprecated, please use ClientTrace/ServerTrace extensions")]
-    public static class TraceHelper
-    {
-        /// <summary>
-        /// Runs the action and adds an error annotation in case of failure
-        /// </summary>
-        /// <param name="action"></param>
-        public static void TracedAction(Action action)
-        {
-            try
-            {
-                action();
-            }
-            catch (Exception ex)
-            {
-                if (Trace.Current != null)
-                {
-                    Trace.Current.Record(Annotations.Tag("error", ex.Message));
-                }
-                throw;
-            }
-        }
-
-        /// <summary>
-        /// Runs the task asynchronously and adds an error annotation in case of failure
-        /// </summary>
-        /// <param name="task"></param>
-        public static async Task TracedActionAsync(Task task)
-        {
-            try
-            {
-                await task;
-            }
-            catch (Exception ex)
-            {
-                if (Trace.Current != null)
-                {
-                    Trace.Current.Record(Annotations.Tag("error", ex.Message));
-                }
-
-                throw;
-            }
-        }
-
-        /// <summary>
-        /// Runs the task asynchronously and adds an error annotation in case of failure
-        /// </summary>
-        /// <param name="task"></param>
-        public static async Task<T> TracedActionAsync<T>(Task<T> task)
-        {
-            try
-            {
-                return await task;
-            }
-            catch (Exception ex)
-            {
-                if (Trace.Current != null)
-                {
-                    Trace.Current.Record(Annotations.Tag("error", ex.Message));
-                }
-
-                throw;
-            }
-        }
-    }
-}
